feat: add -C --camelCaseTypes option
This option adds the ability to camel case just the type names - which
gives a good mix between using JS Standard Camel Case and still
following the actual definitions of database.

The issue with using camel case for both the types and the keys
is that we would have to provide a layer within the programs using the
types to convert back to the original form if the attributes are
different in JS than in the schema.

There are definately
issues with this, especially with a database schema with an inconsistent
naming convention - we would have to provide some sort of mapping file
to acheive correct conversion.

The types on the other hand, only exist in JS and therefore can be named
whatever we want when generating the types.